extract test client to a separate file
